rename mint->stake, mintAndDeposit->stakeAndDeposit, redeem->unstake
